+++ linux/litertlm_binaries.cmake
@@ -1,4 +1,4 @@
 set(LITERTLM_X86_64_URL
-    "https://files.pythonhosted.org/packages/36/27/bb0c2e084d59938bc12f960db50e7e8e056337ed286f31468aaa39bc9d9a/litert_lm_api-0.13.1-py3-none-manylinux_2_27_x86_64.whl")
+    "https://files.pythonhosted.org/packages/7a/22/9f40a16fa74b0410658ec1d7372fd52b5a6a6b9ef5864845632538a756d2/litert_lm_api-0.14.0-py3-none-manylinux_2_27_x86_64.whl")
 set(LITERTLM_ARM64_URL
-    "https://files.pythonhosted.org/packages/12/a7/c23db30dc19ad07a188eeeab48fd8d676554e867dac6e9dbf62be0f82f1/litert_lm_api-0.13.1-py3-none-manylinux_2_27_aarch64.whl")
+    "https://files.pythonhosted.org/packages/f8/41/13da5ec4af33394c8bd1138f0ba87776b5fd6e0ba812493ccde44bbc05c/litert_lm_api-0.14.0-py3-none-manylinux_2_27_aarch64.whl")